rename expect.test to uxCapture.test; one test file per js file
